src/requirements.txt
cachetools==4.1.0
category-encoders==2.2.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
numpy==1.18.4
oauth2client==4.1.3
pandas==1.0.4
patsy==0.5.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
statsmodels==0.11.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
xgboost==0.90
